+++ vector/src/main/java/im/vector/activity/LoginActivity.java
@@ -1996,13 +1996,13 @@ public class LoginActivity extends MXCActionBarActivity implements RegistrationM
         // 2 - the password has been resetted and the user is invited to switch to the login screen
         mForgotValidateEmailButton.setText(mIsPasswordResetted ? R.string.auth_return_to_login : R.string.auth_reset_password_next_step_button);
 
-        @ColorInt final int green = ContextCompat.getColor(this, R.color.vector_green_color);
+        @ColorInt final int accent = ThemeUtils.INSTANCE.getColor(this, R.attr.colorAccent);
         @ColorInt final int white = ContextCompat.getColor(this, android.R.color.white);
 
-        mLoginButton.setBackgroundColor(isLoginMode ? green : white);
-        mLoginButton.setTextColor(!isLoginMode ? green : white);
-        mRegisterButton.setBackgroundColor(!isLoginMode ? green : white);
-        mRegisterButton.setTextColor(isLoginMode ? green : white);
+        mLoginButton.setBackgroundColor(isLoginMode ? accent : white);
+        mLoginButton.setTextColor(!isLoginMode ? accent : white);
+        mRegisterButton.setBackgroundColor(!isLoginMode ? accent : white);
+        mRegisterButton.setTextColor(isLoginMode ? accent : white);
     }
 
     /**