+++ vector/src/main/java/im/vector/activity/MediaPreviewerActivity.java
@@ -15,7 +15,6 @@ import android.widget.VideoView;
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.request.RequestOptions;
 
-import org.jetbrains.annotations.NotNull;
 import org.matrix.androidsdk.data.RoomMediaMessage;
 import org.matrix.androidsdk.util.Log;
 
@@ -25,7 +24,6 @@ import butterknife.BindView;
 import butterknife.OnClick;
 import im.vector.R;
 import im.vector.adapters.MediaPreviewAdapter;
-import im.vector.ui.themes.ActivityOtherThemes;
 
 /**
  * Previews media selected to be send.
@@ -58,12 +56,6 @@ public class MediaPreviewerActivity extends MXCActionBarActivity implements Medi
     @Override
     public int getLayoutRes() {
         return R.layout.activity_media_previewer;
-    }
-
-    @NotNull
-    @Override
-    public ActivityOtherThemes getOtherThemes() {
-        return ActivityOtherThemes.NoActionBar.INSTANCE;
     }
 
     @SuppressLint("ClickableViewAccessibility")