+++ vector/src/main/java/im/vector/activity/StickerPickerActivity.kt
@@ -23,7 +23,6 @@ import android.view.MenuItem
 import im.vector.R
 import im.vector.activity.util.INTEGRATION_MANAGER_ACTIVITY_REQUEST_CODE
 import im.vector.types.JsonDict
-import im.vector.ui.themes.ActivityOtherThemes
 import org.matrix.androidsdk.util.JsonUtils
 import org.matrix.androidsdk.util.Log
 import java.net.URLEncoder
@@ -40,8 +39,6 @@ class StickerPickerActivity : AbstractWidgetActivity() {
     /* ==========================================================================================
      * IMPLEMENT METHODS
      * ========================================================================================== */
-
-    override fun getOtherThemes() = ActivityOtherThemes.NoActionBar
 
     override fun getLayoutRes() = R.layout.activity_choose_sticker