+++ vector/src/main/java/im/vector/activity/VectorRoomActivity.java
@@ -73,7 +73,6 @@ import org.matrix.androidsdk.data.RoomPreviewData;
 import org.matrix.androidsdk.data.RoomState;
 import org.matrix.androidsdk.data.RoomSummary;
 import org.matrix.androidsdk.db.MXLatestChatMessageCache;
-import org.matrix.androidsdk.fragments.IconAndTextDialogFragment;
 import org.matrix.androidsdk.fragments.MatrixMessageListFragment;
 import org.matrix.androidsdk.listeners.IMXNetworkEventListener;
 import org.matrix.androidsdk.listeners.MXEventListener;
@@ -108,6 +107,9 @@ import im.vector.R;
 import im.vector.VectorApp;
 import im.vector.ViewedRoomTracker;
 import im.vector.activity.util.RequestCodesKt;
+import im.vector.dialogs.DialogCallAdapter;
+import im.vector.dialogs.DialogListItem;
+import im.vector.dialogs.DialogSendItemAdapter;
 import im.vector.extensions.MatrixSdkExtensionsKt;
 import im.vector.features.hhs.LimitResourceState;
 import im.vector.features.hhs.ResourceLimitEventListener;
@@ -117,7 +119,6 @@ import im.vector.fragments.VectorUnknownDevicesFragment;
 import im.vector.listeners.IMessagesAdapterActionsListener;
 import im.vector.notifications.NotificationUtils;
 import im.vector.services.EventStreamService;
-import im.vector.ui.themes.ActivityOtherThemes;
 import im.vector.ui.themes.ThemeUtils;
 import im.vector.util.CallsManager;
 import im.vector.util.ExternalApplicationsUtilKt;
@@ -177,8 +178,6 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
     public static final String EXTRA_START_CALL_ID = "EXTRA_START_CALL_ID";
 
     private static final String TAG_FRAGMENT_MATRIX_MESSAGE_LIST = "TAG_FRAGMENT_MATRIX_MESSAGE_LIST";
-    private static final String TAG_FRAGMENT_ATTACHMENTS_DIALOG = "TAG_FRAGMENT_ATTACHMENTS_DIALOG";
-    private static final String TAG_FRAGMENT_CALL_OPTIONS = "TAG_FRAGMENT_CALL_OPTIONS";
 
     private static final String LOG_TAG = VectorRoomActivity.class.getSimpleName();
     private static final int TYPING_TIMEOUT_MS = 10000;
@@ -609,12 +608,6 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
     //================================================================================
     // Activity classes
     //================================================================================
-
-    @NotNull
-    @Override
-    public ActivityOtherThemes getOtherThemes() {
-        return ActivityOtherThemes.NoActionBar.INSTANCE;
-    }
 
     @Override
     public int getLayoutRes() {
@@ -1227,7 +1220,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
             // can only manage one call instance.
             // either there is no active call or resume the active one
             if ((null == call) || call.getCallId().equals(mCallId)) {
-                final Intent intent = new Intent(VectorRoomActivity.this, VectorCallViewActivity.class);
+                final Intent intent = new Intent(this, VectorCallViewActivity.class);
                 intent.putExtra(VectorCallViewActivity.EXTRA_MATRIX_ID, mSession.getCredentials().userId);
                 intent.putExtra(VectorCallViewActivity.EXTRA_CALL_ID, mCallId);
 
@@ -1560,7 +1553,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
                 try {
                     enableActionBarHeader(HIDE_ACTION_BAR_HEADER);
 
-                    final Intent searchIntent = new Intent(VectorRoomActivity.this, VectorUnifiedSearchActivity.class);
+                    final Intent searchIntent = new Intent(this, VectorUnifiedSearchActivity.class);
                     searchIntent.putExtra(VectorUnifiedSearchActivity.EXTRA_ROOM_ID, mRoom.getRoomId());
                     startActivity(searchIntent);
                 } catch (Exception e) {
@@ -1581,7 +1574,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
             case R.id.ic_action_room_leave:
                 if (null != mRoom) {
                     Log.d(LOG_TAG, "Leave the room " + mRoom.getRoomId());
-                    new AlertDialog.Builder(VectorRoomActivity.this)
+                    new AlertDialog.Builder(this)
                             .setTitle(R.string.room_participants_leave_prompt_title)
                             .setMessage(R.string.room_participants_leave_prompt_msg)
                             .setPositiveButton(R.string.leave, new DialogInterface.OnClickListener() {
@@ -1683,7 +1676,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
      */
     private void displayConfCallNotAllowed() {
         // display the dialog with the info text
-        new AlertDialog.Builder(VectorRoomActivity.this)
+        new AlertDialog.Builder(this)
                 .setTitle(R.string.missing_permissions_title_to_start_conf_call)
                 .setMessage(R.string.missing_permissions_to_start_conf_call)
                 .setIcon(android.R.drawable.ic_dialog_alert)
@@ -1700,55 +1693,55 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
         // hide the header room
         enableActionBarHeader(HIDE_ACTION_BAR_HEADER);
 
-        final Integer[] lIcons = new Integer[]{R.drawable.voice_call_green, R.drawable.video_call_green};
-        final Integer[] lTexts = new Integer[]{R.string.action_voice_call, R.string.action_video_call};
+        new AlertDialog.Builder(this)
+                .setAdapter(new DialogCallAdapter(this), new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        onCallItemClicked(which);
+                    }
+                })
+                .setNegativeButton(R.string.cancel, null)
+                .show();
+    }
 
-        IconAndTextDialogFragment fragment = IconAndTextDialogFragment.newInstance(lIcons, lTexts,
-                ThemeUtils.INSTANCE.getColor(this, R.attr.vctr_riot_primary_background_color),
-                ThemeUtils.INSTANCE.getColor(this, R.attr.vctr_riot_primary_text_color));
-        fragment.setOnClickListener(new IconAndTextDialogFragment.OnItemClickListener() {
-            @Override
-            public void onItemClick(IconAndTextDialogFragment dialogFragment, int position) {
-                final boolean isVideoCall;
-                final int permissions;
-                final int requestCode;
+    /**
+     * @param which 0 for voice call, 1 for video call
+     */
+    private void onCallItemClicked(int which) {
+        final boolean isVideoCall;
+        final int permissions;
+        final int requestCode;
 
+        if (which == 0) {
+            isVideoCall = false;
+            permissions = PermissionsToolsKt.PERMISSIONS_FOR_AUDIO_IP_CALL;
+            requestCode = PermissionsToolsKt.PERMISSION_REQUEST_CODE_AUDIO_CALL;
+        } else {
+            isVideoCall = true;
+            permissions = PermissionsToolsKt.PERMISSIONS_FOR_VIDEO_IP_CALL;
+            requestCode = PermissionsToolsKt.PERMISSION_REQUEST_CODE_VIDEO_CALL;
+        }
 
-                if (position == 1) {
-                    isVideoCall = true;
-                    permissions = PermissionsToolsKt.PERMISSIONS_FOR_VIDEO_IP_CALL;
-                    requestCode = PermissionsToolsKt.PERMISSION_REQUEST_CODE_VIDEO_CALL;
-                } else {
-                    isVideoCall = false;
-                    permissions = PermissionsToolsKt.PERMISSIONS_FOR_AUDIO_IP_CALL;
-                    requestCode = PermissionsToolsKt.PERMISSION_REQUEST_CODE_AUDIO_CALL;
-                }
+        AlertDialog.Builder builder = new AlertDialog.Builder(VectorRoomActivity.this)
+                .setTitle(R.string.dialog_title_confirmation);
 
-                AlertDialog.Builder builder = new AlertDialog.Builder(VectorRoomActivity.this)
-                        .setTitle(R.string.dialog_title_confirmation);
+        if (isVideoCall) {
+            builder.setMessage(getString(R.string.start_video_call_prompt_msg));
+        } else {
+            builder.setMessage(getString(R.string.start_voice_call_prompt_msg));
+        }
 
-                if (isVideoCall) {
-                    builder.setMessage(getString(R.string.start_video_call_prompt_msg));
-                } else {
-                    builder.setMessage(getString(R.string.start_voice_call_prompt_msg));
-                }
-
-                builder
-                        .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
-                            @Override
-                            public void onClick(DialogInterface dialog, int which) {
-                                if (PermissionsToolsKt.checkPermissions(permissions, VectorRoomActivity.this, requestCode)) {
-                                    startIpCall(PreferencesManager.useJitsiConfCall(VectorRoomActivity.this), isVideoCall);
-                                }
-                            }
-                        })
-                        .setNegativeButton(R.string.cancel, null)
-                        .show();
-            }
-        });
-
-        // display the fragment dialog
-        fragment.show(getSupportFragmentManager(), TAG_FRAGMENT_CALL_OPTIONS);
+        builder
+                .setPositiveButton(R.string.ok, new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        if (PermissionsToolsKt.checkPermissions(permissions, VectorRoomActivity.this, requestCode)) {
+                            startIpCall(PreferencesManager.useJitsiConfCall(VectorRoomActivity.this), isVideoCall);
+                        }
+                    }
+                })
+                .setNegativeButton(R.string.cancel, null)
+                .show();
     }
 
     /**
@@ -1758,7 +1751,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
      * @param aIsVideoCall true if it is a video call
      */
     private void launchJitsiActivity(Widget widget, boolean aIsVideoCall) {
-        final Intent intent = new Intent(VectorRoomActivity.this, JitsiCallActivity.class);
+        final Intent intent = new Intent(this, JitsiCallActivity.class);
         intent.putExtra(JitsiCallActivity.EXTRA_WIDGET_ID, widget);
         intent.putExtra(JitsiCallActivity.EXTRA_ENABLE_VIDEO, aIsVideoCall);
         startActivity(intent);
@@ -2175,7 +2168,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
 
         final boolean typingStatus = isTyping;
 
-        mRoom.sendTypingNotification(typingStatus, notificationTimeoutMS, new SimpleApiCallback<Void>(VectorRoomActivity.this) {
+        mRoom.sendTypingNotification(typingStatus, notificationTimeoutMS, new SimpleApiCallback<Void>(this) {
             @Override
             public void onSuccess(Void info) {
                 // Reset last typing date
@@ -2239,7 +2232,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
             enableActionBarHeader(HIDE_ACTION_BAR_HEADER);
 
             // pop to the home activity
-            Intent intent = new Intent(VectorRoomActivity.this, VectorRoomDetailsActivity.class);
+            Intent intent = new Intent(this, VectorRoomDetailsActivity.class);
             intent.putExtra(VectorRoomDetailsActivity.EXTRA_ROOM_ID, mRoom.getRoomId());
             intent.putExtra(VectorRoomDetailsActivity.EXTRA_MATRIX_ID, mSession.getCredentials().userId);
             intent.putExtra(VectorRoomDetailsActivity.EXTRA_SELECTED_TAB_ID, selectedTab);
@@ -3515,7 +3508,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
             }
 
             // clear the activity stack to home activity
-            Intent intent = new Intent(VectorRoomActivity.this, VectorHomeActivity.class);
+            Intent intent = new Intent(this, VectorHomeActivity.class);
             intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_SINGLE_TOP | Intent.FLAG_ACTIVITY_NEW_TASK);
 
             intent.putExtra(VectorHomeActivity.EXTRA_JUMP_TO_ROOM_PARAMS, (HashMap) params);
@@ -3942,77 +3935,61 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
             // hide the header room
             enableActionBarHeader(HIDE_ACTION_BAR_HEADER);
 
-            FragmentManager fm = getSupportFragmentManager();
-            IconAndTextDialogFragment fragment = (IconAndTextDialogFragment) fm.findFragmentByTag(TAG_FRAGMENT_ATTACHMENTS_DIALOG);
-
-            if (fragment != null) {
-                fragment.dismissAllowingStateLoss();
-            }
-
-            List<Integer> messagesList = new ArrayList<>();
-            List<Integer> iconsList = new ArrayList<>();
+            final List<DialogListItem> items = new ArrayList<>();
 
             // Send file
-            messagesList.add(R.string.option_send_files);
-            iconsList.add(R.drawable.ic_material_file);
+            items.add(DialogListItem.SendFile.INSTANCE);
 
             // Send voice
-            if (PreferencesManager.isSendVoiceFeatureEnabled(VectorRoomActivity.this)) {
-                messagesList.add(R.string.option_send_voice);
-                iconsList.add(R.drawable.vector_micro_green);
+            if (PreferencesManager.isSendVoiceFeatureEnabled(this)) {
+                items.add(DialogListItem.SendVoice.INSTANCE);
             }
 
             // Send sticker
-            messagesList.add(R.string.option_send_sticker);
-            iconsList.add(R.drawable.ic_send_sticker);
+            items.add(DialogListItem.SendSticker.INSTANCE);
 
             // Camera
-            if (PreferencesManager.useNativeCamera(VectorRoomActivity.this)) {
-                messagesList.add(R.string.option_take_photo);
-                iconsList.add(R.drawable.ic_material_camera);
-                messagesList.add(R.string.option_take_video);
-                iconsList.add(R.drawable.ic_material_videocam);
+            if (PreferencesManager.useNativeCamera(this)) {
+                items.add(DialogListItem.TakePhoto.INSTANCE);
+                items.add(DialogListItem.TakeVideo.INSTANCE);
             } else {
-                messagesList.add(R.string.option_take_photo_video);
-                iconsList.add(R.drawable.ic_material_camera);
+                items.add(DialogListItem.TakePhotoVideo.INSTANCE);
             }
 
-            final Integer[] messages = messagesList.toArray(new Integer[0]);
-            final Integer[] icons = iconsList.toArray(new Integer[0]);
+            new AlertDialog.Builder(this)
+                    .setAdapter(new DialogSendItemAdapter(this, items), new DialogInterface.OnClickListener() {
+                        @Override
+                        public void onClick(DialogInterface dialog, int which) {
+                            onSendChoiceClicked(items.get(which));
+                        }
+                    })
+                    .setNegativeButton(R.string.cancel, null)
+                    .show();
+        }
+    }
 
-            fragment = IconAndTextDialogFragment.newInstance(icons, messages,
-                    ThemeUtils.INSTANCE.getColor(VectorRoomActivity.this, R.attr.vctr_riot_primary_background_color),
-                    ThemeUtils.INSTANCE.getColor(VectorRoomActivity.this, R.attr.vctr_riot_primary_text_color));
-            fragment.setOnClickListener(new IconAndTextDialogFragment.OnItemClickListener() {
-                @Override
-                public void onItemClick(IconAndTextDialogFragment dialogFragment, int position) {
-                    Integer selectedVal = messages[position];
-
-                    if (selectedVal == R.string.option_send_files) {
-                        launchFileSelectionIntent();
-                    } else if (selectedVal == R.string.option_send_voice) {
-                        launchAudioRecorderIntent();
-                    } else if (selectedVal == R.string.option_send_sticker) {
-                        startStickerPickerActivity();
-                    } else if (selectedVal == R.string.option_take_photo_video) {
-                        if (PermissionsToolsKt.checkPermissions(PermissionsToolsKt.PERMISSIONS_FOR_TAKING_PHOTO,
-                                VectorRoomActivity.this, PermissionsToolsKt.PERMISSION_REQUEST_CODE_LAUNCH_CAMERA)) {
-                            launchCamera();
-                        }
-                    } else if (selectedVal == R.string.option_take_photo) {
-                        if (PermissionsToolsKt.checkPermissions(PermissionsToolsKt.PERMISSIONS_FOR_TAKING_PHOTO,
-                                VectorRoomActivity.this, PermissionsToolsKt.PERMISSION_REQUEST_CODE_LAUNCH_NATIVE_CAMERA)) {
-                            launchNativeCamera();
-                        }
-                    } else if (selectedVal == R.string.option_take_video) {
-                        if (PermissionsToolsKt.checkPermissions(PermissionsToolsKt.PERMISSIONS_FOR_TAKING_PHOTO,
-                                VectorRoomActivity.this, PermissionsToolsKt.PERMISSION_REQUEST_CODE_LAUNCH_NATIVE_VIDEO_CAMERA)) {
-                            launchNativeVideoRecorder();
-                        }
-                    }
-                }
-            });
-            fragment.show(fm, TAG_FRAGMENT_ATTACHMENTS_DIALOG);
+    private void onSendChoiceClicked(DialogListItem dialogListItem) {
+        if (dialogListItem instanceof DialogListItem.SendFile) {
+            launchFileSelectionIntent();
+        } else if (dialogListItem instanceof DialogListItem.SendVoice) {
+            launchAudioRecorderIntent();
+        } else if (dialogListItem instanceof DialogListItem.SendSticker) {
+            startStickerPickerActivity();
+        } else if (dialogListItem instanceof DialogListItem.TakePhotoVideo) {
+            if (PermissionsToolsKt.checkPermissions(PermissionsToolsKt.PERMISSIONS_FOR_TAKING_PHOTO,
+                    VectorRoomActivity.this, PermissionsToolsKt.PERMISSION_REQUEST_CODE_LAUNCH_CAMERA)) {
+                launchCamera();
+            }
+        } else if (dialogListItem instanceof DialogListItem.TakePhoto) {
+            if (PermissionsToolsKt.checkPermissions(PermissionsToolsKt.PERMISSIONS_FOR_TAKING_PHOTO,
+                    VectorRoomActivity.this, PermissionsToolsKt.PERMISSION_REQUEST_CODE_LAUNCH_NATIVE_CAMERA)) {
+                launchNativeCamera();
+            }
+        } else if (dialogListItem instanceof DialogListItem.TakeVideo) {
+            if (PermissionsToolsKt.checkPermissions(PermissionsToolsKt.PERMISSIONS_FOR_TAKING_PHOTO,
+                    VectorRoomActivity.this, PermissionsToolsKt.PERMISSION_REQUEST_CODE_LAUNCH_NATIVE_VIDEO_CAMERA)) {
+                launchNativeVideoRecorder();
+            }
         }
     }
 
@@ -4020,7 +3997,7 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
     void onPendingCallClick() {
         IMXCall call = CallsManager.getSharedInstance().getActiveCall();
         if (null != call) {
-            final Intent intent = new Intent(VectorRoomActivity.this, VectorCallViewActivity.class);
+            final Intent intent = new Intent(this, VectorCallViewActivity.class);
             intent.putExtra(VectorCallViewActivity.EXTRA_MATRIX_ID, call.getSession().getCredentials().userId);
             intent.putExtra(VectorCallViewActivity.EXTRA_CALL_ID, call.getCallId());
             startActivity(intent);
@@ -4045,14 +4022,14 @@ public class VectorRoomActivity extends MXCActionBarActivity implements
     void onStartCallClick() {
         if ((null != mRoom) && mRoom.isEncrypted() && (mRoom.getNumberOfMembers() > 2)) {
             // display the dialog with the info text
-            new AlertDialog.Builder(VectorRoomActivity.this)
+            new AlertDialog.Builder(this)
                     .setMessage(R.string.room_no_conference_call_in_encrypted_rooms)
                     .setIcon(android.R.drawable.ic_dialog_alert)
                     .setPositiveButton(R.string.ok, null)
                     .show();
         } else if (isUserAllowedToStartConfCall()) {
             if (mRoom.getNumberOfMembers() > 2) {
-                new AlertDialog.Builder(VectorRoomActivity.this)
+                new AlertDialog.Builder(this)
                         .setTitle(R.string.conference_call_warning_title)
                         .setMessage(R.string.conference_call_warning_message)
                         .setIcon(android.R.drawable.ic_dialog_alert)