+++ vector/src/main/java/im/vector/activity/VectorSettingsActivity.kt
@@ -37,6 +37,8 @@ class VectorSettingsActivity : MXCActionBarActivity() {
     }
 
     override fun initUiAndData() {
+        configureToolbar()
+
         var session = getSession(intent)
 
         if (null == session) {