+++ vector/src/main/java/im/vector/dialogs/DialogCallAdapter.kt
@@ -1,6 +1,4 @@
 /*
- * Copyright 2014 OpenMarket Ltd
- * Copyright 2017 Vector Creations Ltd
  * Copyright 2018 New Vector Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -16,21 +14,14 @@
  * limitations under the License.
  */
 
-package im.vector.activity;
+package im.vector.dialogs
 
-import im.vector.R;
+import android.content.Context
 
-/**
- * LoggingOutActivity displays an animation while a session log out is in progress.
- */
-public class LoggingOutActivity extends MXCActionBarActivity {
-    @Override
-    public int getLayoutRes() {
-        return R.layout.vector_activity_splash;
-    }
+internal class DialogCallAdapter(context: Context) : DialogAdapter(context) {
 
-    @Override
-    public void initUiAndData() {
-        // Nothing to do
+    init {
+        add(DialogListItem.StartVoiceCall)
+        add(DialogListItem.StartVideoCall)
     }
 }